requirements.txt
# پردازش متن فارسی
hazm
python-docx

# Embedding و Vector DB
requests
chromadb
numpy

# تحلیل متن (برای Stage 2)
scikit-learn

# تست
pytest
pytest-cov

# ابزار
python-dotenv